Include maven in this project
