--- src/com/dostojic/climbers/exception/LoginException.java
@@ -1,33 +1,0 @@
-/*
- * To change this license header, choose License Headers in Project Properties.
- * To change this template file, choose Tools | Templates
- * and open the template in the editor.
- */
-package com.dostojic.climbers.exception;
-
-/**
- *
- * @author planina
- */
-public class LoginException extends Exception {
-
-    public LoginException() {
-    }
-
-    public LoginException(String message) {
-        super(message);
-    }
-
-    public LoginException(String message, Throwable cause) {
-        super(message, cause);
-    }
-
-    public LoginException(Throwable cause) {
-        super(cause);
-    }
-
-    public LoginException(String message, Throwable cause, boolean enableSuppression, boolean writableStackTrace) {
-        super(message, cause, enableSuppression, writableStackTrace);
-    }
-
-}
